Font Library: refactor FontsGrid component
